seperate xcomponent call from ui
